+++ duckduckgo/__init__.py
@@ -1,0 +1,11 @@
+# duckduckgo.py - Library for querying the DuckDuckGo API
+#
+# Copyright (c) 2010 Michael Stephens <me@mikej.st>
+# Copyright (c) 2012-2013 Michael Smith <crazedpsyc@gshellz.org>
+# Copyright (c) 2017 Members of the Programming Server
+#
+# See LICENSE for terms of usage, modification and redistribution.
+
+__version__ = 0.300
+
+from .query import query, get_zci
